pch include 정리 -> 컴파일 시간 단축

+++ DX12RenderingDemo/Timer.h
@@ -25,7 +25,7 @@ private:
     bool hardwareHasPerformanceCounter_ = false;
     bool stopped_ = false;
 
-    float timeScale_ = 0.0f;
+    double timeScale_ = 0.0f;
     float timeElapsed_ = 0.0f;
 
     __int64 baseTime_ = 0;
